Fix concurrent generator use
Previously running concurrency operations against the generator would
result in a panic because it was actually taking maps that it found and
fixtures and mutating them, resulting in concurrent reads and writes.

This patch introduces a fix that copies fixtures that are found in the
decoded set so that each Goroutine is working with a local copy. It also
introduces testing to protect against regressions.

Fixes #10.