Coerce integer-indexed maps to slices
This is unusual for normal "Rack-style" form encoding, but at Stripe on
some endpoints we allow maps that are indexed with integers to be
changed to arrays in the backend. So for example this:

    myarr[0]=foo

Would be interpreted in the API as:

    ["foo"]

The purpose of this is to be able to use arbitrary indexing to update
only certain items, so this also works:

    myarr[3]=foo

Becomes:

    [nil, nil, nil "foo"]

You can find an example of this in the `items` parameter on create
subscription.

Here we augment the coercer to detect if the target type is `array` and
then try to coerce a map to a slice as long as all its indexes were
strings that looked like integers.

I've added a test suite, but have also verified that this fixes a
failing test in my stripe-go PR that's targeting stripe-mock.