Descend into arrays while doing parameter coercing
This patch adds logic to the coercer so that if it runs into an array,
it can descend into it and coerce its elements according to the JSON
schema defined in `items`.

I found that I need this to get stripe-go's test suite running with
respect to setting a subscription's items.