Fix decoded maps when unmarshaling fixtures from a YAML file
As shown in #22, when we try to unmarshal fixtures from a YAML file, we
end up with a crash in the generator due to a type mismatch.

This turns out to be because of [1]. Because YAML technically supports
boolean keys, the YAML library unmarshals to maps in a way to handle
them (`map[interface{}]interface{}`), even if it's not suitable for most
users (the issue strongly suggests that a lot of people would like
`map[string]interface{}`).

Here we work around the problem by post-processing fixtures that we've
loaded via YAML by recursing through and changing all instances of
`map[interface{}]interface{}` to `map[string]interface{}`. I've added a
few basic tests, but also tested that the call reported in #22 now works
as expected.

Fixes #22.

[1] https://github.com/go-yaml/yaml/issues/139