Send valid path to generator while running test suite
So I couldn't figure out what exactly changed here, but as reported in
our channel yesterday, `stripe-mock`'s test suite fails on the latest
version of the OpenAPI spec.

The reason is that when generating list resources, we're currently
filling in `url` with the dud string `<test request path>`, and doing so
causes the generated resource to fail to validate because that string is
not in the OpenAPI spec's `enum` set for `url`.

Here we resolve the problem by plumbing through each URL's path from the
ingested OpenAPI spec and sending it onto the generator.

This doesn't quite fix the problems with the latest OpenAPI though --
there's at least one more case of an invalid fixture which I'm trying to
resolve separately.