Special case more non-RESTful "actions"
This addresses half of #61 by special-casing more non-RESTful "actions"
that are known to exist in the API. This will allow primary object IDs
in URLs for those actions to be reflected into the final object
(currently they don't get caught by the matcher and are not).

I'm planning on rebuilding this to handle extraction of path parameters
more elegantly, but this will get us to the 95% mark for now.

See also #61.