Fix non-deterministic test
I recently introduced a test that checks its results against a value
generated by iterating a map. In Go, the order with which a map is
iterated is not guaranteed, and so the test will occasionally fail as
the map is iterated in an order contrary to its insertion order.

This patch sorts the output slice before joining it, thereby fixing the
non-determinism. This is a case that's only occasionally hit and only in
verbose mode, so the performance implications are miniscule.