Implement "advanced" path parameter response reflection
Builds on reflecting URL parameters as introduced in #59 to also support
reflecting parameters beyond just the object's primary ID. More
concretely, if we have a path like this:

    /application_fees/fee_123/refunds/re_123

Before we'd just reflect the last part in (`re_123`), but here we now
reflect the parent object in as well (`fee_123`).

The one major caveat to this is that currently our query parameters are
not very introspectable in that we don't really know what kind of object
`fee_123` is supposed to be. To that end, we use the name of the
parameter (in this case `fee`) and compare that it against two places
to see if it's likely a referencing this object:

    * A fixture's `object` value.
    * The name of the key containing an ID value of subobject.

This will work for some objects, but I'm probably going to go through
and either (1) do a pass in the backend to correctly name parameters, or
(2) implement a lookup table in `stripe-mock` to get this working 100%.