Flexible `AdditionalProperties`
Previously, we only ever had the boolean use of `AdditionalProperties`,
in that a `false` was used to specify that no additional properties were
allowed.

As of more recent versions of the spec, we're now using the schema
format for it as well, and this causes the YAML parser to error as it
encounters a value that's not compatible with the `bool` that we had
specified before.

This patch changes `AdditionalProperties` to take an `interface{}` so
that either type can be accepted. This won't be workable if we actually
need a value for `AdditionalProperties`, but we're not using this
property at the moment, so it's fine for now.