Validate `multipart/form-data` + error on mismatched media types
This makes the changes necessary to be able to successfully validate
payloads sent as `multipart/form-data` along with the standard
`application/x-www-form-urlencoded`.

To accomplish this, I ended up splitting the parameter parser into two
separate components:

1. A new "parser" that just splits an incoming string into form
   key/value tuples. This is used to parse query strings and body form
   data. We prefer it over Go's built-in parser because it returns
   an ordered slice instead of an unordered map, which is critical for
   correctly parsing incoming arrays and maps.
2. An "assembler" that takes the parser's output and assembles
   higher-level data structures from it, including arrays and maps by
   using the "Rack-style" conventions in use by the Stripe API.

To parse `multipart/form-data`, we rely on Go's built-in multipart
parser, and then hands off its results to our new assembler. You can see
all of this at work in `param.ParseParams`.

In addition, it tweaks behavior so that if a client sends either an
empty `Content-Type` header where one was expected (i.e., usually
non-`GET`) or a `Content-Type` that doesn't match the expected media
type of the responding endpoint, `stripe-mock` will respond with an
error.

This last part is intended to be an extra safety feature to help ensure
that client implementations are correct. In particular, it should be
helpful in making sure that file API endpoints receive
`multipart/form-data` instead of the standard
`application/x-www-form-urlencoded`.

So the one big caveat is that file uploads in `stripe-ruby` and other
libraries will still not quite succeed against these changes. To get
them fully working we need to change the type of the "file create"
endpoint's `file` parameter from an object to a string. I have another
pull going into the backend to do this.

Fixes #69.