Allow empty content type on `DELETE`
We recently added checks in `stripe-mock` to make sure that the right
`Content-Type` was being sent. Unfortunately, the logic was not quite
right for deletes, where you occasionally see a `Content-Type` in rare
cases, but which usually have empty request bodies, and therefore an
empty `Content-Type` is allowed.

This patch brings in two fixes:

1. The OpenAPI spec has been tweak so as to not incorrectly include
   `expand` parameters on `DELETE` operations. This means that most
   `DELETE` operations have no defined request schema at all.
2. Special cases incoming `DELETE` operations so that they can be
   checked against a schema if they provide parameters, but are also
   allowed to specify an empty `Content-Type` as well.

Between the two improvements, all edges should be covered.

The impetus here is to resolve [1] where a `stripe-mock` upgrade is
incompatible with the current code. I've verified that with this patch
included, `stripe-python`'s test suite runs successfully.

[1] https://github.com/stripe/stripe-python/pull/426