Replace data in response with data from request
Here we try to to some additional enrichment on response data by
replacing what we found in fixtures with data that came in from the
request.

The algorithm is quite simple: we walk the decoded request object and
the generated response object, replace any scalars that we see where the
key name is the same, and recurse in when we see two keys with the same
names that are both maps.

In addition, I realized that we are no longer vetting that incorrect
keys are not being passed to stripe-mock, which appears to be a
regression from a change passed some time ago. This patch also adds back
a check on `additionalProperties: false`.