Hedge against the "zero Value"
Reflect in Go has the concept of a "zero Value" [1] (not be confused with a
type's zero value with a lowercase "v") and asking for Type on one will
panic. I'm not exactly sure under what conditions these are generated,
but they are occasionally, so here we hedge against them.

See also: https://github.com/stripe/stripe-mock/issues/75

Fixes #75.

[1] https://golang.org/pkg/reflect/#Value