Support simultaneous HTTP and HTTPS
I was a little hasty in adding support for HTTPS before in that when
we're developing with stripe-mock locally, we need to have a setup that
supports both HTTP *and* HTTPS because some libraries will be using one
and some libraries will be using the other.

I've added a number of command line options that allow stripe-mock to be
booted so that it's listening for both HTTP and HTTPS (and HTTP/2)
simultaneously. It's added a little bit of complication, so I've added a
number of checks for incompatible options and more documentation to the
README on how to use them all.

`12111` will remain our default port for HTTP, and `12112` will become
our default port for HTTPS. Libraries that want HTTP will lock to
`12111` by default and those that need HTTPS will lock to `12112`.

See the README for full details, but here are some basic usages with the
options:

``` sh
stripe-mock

stripe-mock -https

stripe-mock -port 80

stripe-mock -https -unix /tmp/stripe-mock-secure.sock

stripe-mock -http-port 80 -https-port 443
```