Decided between returning a deleted versus non-deleted from HTTP method
Implements a simple heuristic so that stripe-mock will pick an `anyOf`
branch based off whether the incoming HTTP request method is a `DELETE`
or or not. If it is `DELETE`, we try to pick the deleted form of a
resource, and vice versa.

We're doing this because we'd occasionally run into a place where
stripe-mock was picking a branch based off its naive "pick the first
branch" logic when it was quite obvious that another branch would
overwhelmingly have been preferred. For example, returning a deleted
customer from `GET /v1/customers/:id` because the deleted customer
appeared first in the `anyOf` list.

The heuristic isn't perfect, but it will usually do the right thing.