Don't replace data on non-`POST` endpoints
stripe-mock has a feature where it reflects data from input parameters
back into the response so that when creating an updating objects, it
produces responses that look more realistic with some of the actual
input values that were sent in.

Unfortunately, we found that this was working a little _too_ well and
being applied to other operations like listing as well. So for example,
if we request a list with a filter of `object=source`, we'd see a
response where the response list object's `object` field was set to
`source` instead of `list` like it should be.

Here we make sure that data replacement only occurs on `POST`
operations and add some tests to verify this behavior.