Coerce values within an object within an `anyOf`
As described in #95, there's currently a problem with the coercer in
that if it finds an `anyOf`, it assumes that only primitive schema types
like a `boolean` or `integer` will be present in it.

This isn't true -- it somes cases (say `destination` while creating a
charge), it's possible to have an object _or_ a primitive type. In the
case of the object, stripe-mock wouldn't error, but would also not
coerce anything within it, and therefore requests might fail validation.

Here we consider that an object might be found as welll, and implement
code to handle that case.

Fixes #95.