Fix test to match naming + less aggressive parameter coercion
Remi saw on #97 that I'd renamed a field to `bool_or_int_key` intending
to change a test, but I'd accidentally forgotten to make that subsequent
change, which was to create a higher fidelity test by testing two
primitive type schemas within the same `anyOf`.

Here I made that change, but noticed when I did it that the tests
started to fail. It turned out this is because we were coercing
primitive types overly aggressively. If for example a boolean schema
received an integer like `123`, `strconv.ParseBool` would fail with an
error, but we'd just default to the value `false`.

In this test, it had the effect of the `anyOf` taking the first boolean
branch, even though the value is much better suited to be coerced to an
integer.

Here we relax primitive type coercion. If a value can be parsed for that
type, it is, but if it's not, we default to the value that was given to
us. This seems like a better fit for the coercion behave that we
actually want.