Listen for both HTTP and HTTPS if no args given
This tweaks stripe-mock's behavior slightly so that it listens for both
HTTP and HTTPS if it's invoked with no arguments. The impetus is we'll
often want HTTPS activated to test *something*, and having it active
costs very little extra (just one additional port and listener, and
those are cheap). This is not a breaking change because the HTTP port we
listen on remains unchanged.

We continue to operate as before when arguments are passed. For example:

* `stripe-mock -http` activates *just* HTTP.
* `stripe-mock -https` activates *just* HTTPS.
* `stripe-mock -http -https` continues to activate both HTTP and HTTPS.

We also plumb a protocol name through to our various listening
functions. This has the effect of producing more informational messages
on startup:

    Listening for HTTP on port: 12111
    Listening for HTTPS on port: 12112

Where before it was just:

    Listening on port: 12111
    Listening on port: 12112