Improve test setup
Previously the listener tests depended on a hard-coded port that we
hoped was free. Here we change the strategy slightly to ask the OS for a
free port with `:0`.

Unfortunately, that means also changing some code elsewhere because we
were previously relying on `0` to signify an empty argument. We now
change that to `-1` and update code and tests.