Add support for query parameters on `GET` endpoints
Add support for validating query parameters on `GET` endpoints
(Non-`GET` endpoints always add all their parameters into the request
body schema in OpenAPI, so we don't have to worry about supporting
them.)

A little bit of background: when generating OpenAPI, we use a simplified
logic that puts parameters for `GET` endpoints in as query parameters,
and puts the parameters for all other types of endpoint in as request
body parameters. This is not exactly right because all endpoints will
read either type of parameter, but it makes things look a little more
correct because convention is to always use query strings on `GET`
endpoints (e.g., `GET /v1/charges?limit=3`).

In stripe-mock, I initially skipped over implementing any kind of
validation on query string parameters because they're not as important
to validate, and it was more code to write. Unfortunately, that
omission carried through all the way to the present day.

When I put stronger extra parameter validation into the OpenAPI spec, it
revealed a problem where stripe-mock was refusing query string
parameters on `GET` endpoints because they weren't in the request body
schema, and it didn't know how to handle the query string parameter
definitions. This is why stripe-ruby's test suite fails against the
latest version of stripe-mock.

Here I add support for query parameters using a simple method that
converts their OpenAPI definitions into a JSON schema so we can use a
similar method to validate them as we do for request bodies.

So here are the basic rules:

1. In OpenAPI `GET` endpoints only specify query parameters.
2. In OpenAPI, non-`GET` endpoints only specify request body schemas.
3. When validating a `GET`, only the query string is considered (an HTTP
   RFC suggests that request bodies on `GET` are semantically
   meaningless, so this is consistent with that, but more importantly, I
   think it'll be a useful simplification for our purposes).
4. When validating non-`GET`, both the query string and request body are
   parsed.

I suspect there might still be some other subtle problem that I haven't
considered yet, but this seems to be a big improvement over what we have
today. I'm able to run the entire stripe-ruby test suite against it with
no failures, and it even caught a few problems in stripe-ruby's test
suite [1].

[1] https://github.com/stripe/stripe-ruby/pull/679