Don't start `HTTPS` with `-port` option specified
I recently did a lot of refactoring for the arguments so that we start
both HTTP and HTTPS by default.

One place where we don't start both of them is if either HTTP or HTTPS
is explicitly requested but the other is not.

This mostly worked, but I forgot to account for the old `-port` option,
which is taken to be an alias for `-http-port`. Like with `-http-port`,
if it's specified but HTTPS is not requested, we start HTTP only. This
patch now accounts for that.