Add new option `-strict-version-check`
Adds a new option called `-strict-version-check` that causes stripe-mock
to check that any value passed in an incoming request's `Stripe-Version`
header matches the version specified in the bundled OpenAPI
specification.

This is intended to act as an optionally strengthened expectation which
protects against an unintended drift between a bundled stripe-mock /
OpenAPI and the version being used from an API library. We'll likely
want to enable the option for static libraries like stripe-go and
stripe-dotnet.