Bring back patch to enrich additionalProperties error messages
This is bad and I should feel bad, but at some point I added a patch to
`jsval` that gives a more detailed error message and the tests depend on
it (it also improves the user experience a lot).

I'm adding this back after updating dependencies, but the right move
here is to try and get this in upstream or fork the repository.