Generates new primary IDs from "create" endpoints
Here we extend the ID replacement infrastructure to generate new IDs
from "create" endpoints. For example:

    $  curl -X POST -H "Content-Type: application/x-www-form-urlencoded" -H "Authorization: Bearer sk_test_123" http://localhost:2345/v1/charges | jq '.id'
    "ch_CNT2UVXcbLq0Jt5"

    $  curl -X POST -H "Content-Type: application/x-www-form-urlencoded" -H "Authorization: Bearer sk_test_123" http://localhost:2345/v1/charges | jq '.id'
    "ch_CNT2bb9wsZpgsxT"

With the idea being that much like you get with the Stripe API, new
objects have unique IDs every time.

I ported the same algorithm we use for token generation internally so
that IDs are made up of a prefix (determined from the source fixture), a
time portion, and a random portion. This means that they order in a
mostly ascending way over time (like the Stripe API this is not
guaranteed but is often true), which is always kind of nice.

Fixes #152.