Handle more special characters and spaces in path parameters
Handles all special characters in path parameters that are defined in
RFC3986 instead of just the couple that we'd special cased before, along
with `+` which represents a space.

After matching a patch, we now also run all parameters through
`url.QueryUnescape` to decode `%`-encoded bytes and convert `+`s back to
spaces.