Deduplicate slashes in incoming paths
Occasionally an API library or other client might do some path appending
that's slightly off and end up with an extra slash in their target.
`//v1/charges` for example.

While the Stripe API will happily accept this, stripe-mock currently
won't because Go intercepts the doubled slashes fairly high up in the
stack and responds with a 301 to a normalized URL (e.g. `/v1/charges`).

Go's default behavior is probably desirable most of the time, but given
that the Stripe API takes doubled slashes, in this case it makes more
sense to accept them as well. This patch introduces a wrapping handler
that deduplicates slashes on incoming requests, thereby allowing a
request with doubled slashes to succeed normally.

Fixes #170.