Rebuild data replacement strategy to use input from OpenAPI schema
Previously, the data replacer worked by taking input request and output
response data, and replacing values if the types matched up. This
generally worked, but would result in some inappropriate replacements
for more complex data types like arrays, as seen in [1].

In this patch we upgrade the replacement strategy so that we perform
replacements if the type of incoming request data matches what we
expected in the response based on the OpenAPI schema. This allows us,
for example, to not only check that a type is an array, but also what
sort of elements that array is supposed to contain, even if the value
from our fixtures is an empty array (which is often the case).

Fixes #210.

[1] https://github.com/stripe/stripe-mock/issues/210